Switch request system to use 'fetch'
We were polyfilling a node module interface into the browser. This made depending on lib-interfaces require a special webpack config... its become harder to manage that. So... we polyfilled fetch into node in the web-service. And are now switching our XHR system to call fetch instead of "request" which built XHR objects. For browsers that do not have fetch natively, we will link-in a polyfill.